recipe_page 파이어베이스에서 읽기완료

+++ app/src/main/java/com/example/cooclock/recipeStepItem.java
@@ -26,11 +26,11 @@ public class recipeStepItem {
     }
 
     public int getMinute() {
-        return Integer.parseInt(time);
+        return (int) Double.parseDouble(time);
     }
 
     public int getSecond() {
-        return (int) ((Double.parseDouble(time)-Integer.parseInt(time))*100);
+        return (int) ((Double.parseDouble(time)-(int) Double.parseDouble(time))*100);
     }
 
     public int getResId() {
